Add and Struct folder monan, fix a little logic in LoaiMonAn
