+++ src/main/java/com/thefourrestaurant/view/monan/GiaoDienChiTietMonAn.java
@@ -1,0 +1,9 @@
+package com.thefourrestaurant.view.monan;
+
+import javafx.scene.layout.VBox;
+
+public class GiaoDienChiTietMonAn extends VBox {
+    public GiaoDienChiTietMonAn() {
+        // Nội dung chi tiết món ăn sẽ được thêm vào đây
+    }
+}
